Stage2 refactored to simplify multiple implementations
