+++ benchmark/parse.cpp
@@ -157,7 +157,8 @@ int main(int argc, char *argv[]) {
       break;
     }
     unified.start();
-    isok = isok && (simdjson::SUCCESS == unified_machine(p.data(), p.size(), pj));
+    // The default template is simdjson::instruction_set::native.
+    isok = isok && (simdjson::SUCCESS == unified_machine<>(p.data(), p.size(), pj));
     unified.end(results);
     cy2 += results[0];
     cl2 += results[1];
@@ -188,7 +189,7 @@ int main(int argc, char *argv[]) {
     auto start = std::chrono::steady_clock::now();
     // The default template is simdjson::instruction_set::native.
     isok = (find_structural_bits<>(p.data(), p.size(), pj) == simdjson::SUCCESS);
-    isok = isok && (simdjson::SUCCESS == unified_machine(p.data(), p.size(), pj));
+    isok = isok && (simdjson::SUCCESS == unified_machine<>(p.data(), p.size(), pj));
     auto end = std::chrono::steady_clock::now();
     std::chrono::duration<double> secs = end - start;
     res[i] = secs.count();
