+++ tests/stringparsingcheck.cpp
@@ -203,7 +203,7 @@ static bool parse_string(const char *p, char *output, char **end) {
 // end of borrowed code
 char *bigbuffer; // global variable
 
-inline void foundBadString(const uint8_t *buf) {
+void foundBadString(const uint8_t *buf) {
   bad_string++;
   char *end;
   if (parse_string((const char *)buf, bigbuffer, &end)) {
@@ -226,7 +226,7 @@ void print_cmp_hex(const char *s1, const char *s2, size_t len) {
   }
 }
 
-inline void foundString(const uint8_t *buf, const uint8_t *parsed_begin,
+void foundString(const uint8_t *buf, const uint8_t *parsed_begin,
                         const uint8_t *parsed_end) {
   size_t thislen = parsed_end - parsed_begin;
   total_string_length += thislen;
